+++ cxroots/contours/annulus_sector.py
@@ -1,4 +1,5 @@
 from math import pi
+from typing import Literal
 
 import numpy as np
 
@@ -44,9 +45,10 @@ class AnnulusSector(Contour):
         annulusSector.show()
     """
 
+    axis_names = ("r", "phi")
+
     def __init__(self, center, radii, phi_range):
         self.center = center
-        self.axis_name = ("r", "phi")
 
         if phi_range[0] > phi_range[1]:
             phi_range = (phi_range[0], phi_range[1] + 2 * pi)
@@ -109,7 +111,7 @@ class AnnulusSector(Contour):
 
         return radius_correct and angle_correct
 
-    def subdivide(self, axis, division_factor=0.5):
+    def subdivide(self, axis: Literal["r", "phi"], division_factor: float = 0.5):
         """
         Subdivide the contour
 
@@ -136,7 +138,8 @@ class AnnulusSector(Contour):
         """
         r0, r1 = self.radii
         phi0, phi1 = self.phi_range
-        if axis == 0 or axis == self.axis_name[0]:
+
+        if axis == "r":
             division_point = r0 + division_factor * (r1 - r0)
             box1 = AnnulusSector(self.center, [r0, division_point], self.phi_range)
             box2 = AnnulusSector(self.center, [division_point, r1], self.phi_range)
@@ -145,22 +148,25 @@ class AnnulusSector(Contour):
             # this allows the cached integrals to be used
             box1.segments[3] = self.segments[3]
             box2.segments[1] = self.segments[1]
-            box1.segments[1]._reversePath = box2.segments[3]
-            box2.segments[3]._reversePath = box1.segments[1]
+            box1.segments[1]._reverse_path = box2.segments[3]
+            box2.segments[3]._reverse_path = box1.segments[1]
 
-        elif axis == 1 or axis == self.axis_name[1]:
+        elif axis == "phi":
             division_point = phi0 + division_factor * (phi1 - phi0)
             box1 = AnnulusSector(self.center, self.radii, [phi0, division_point])
             box2 = AnnulusSector(self.center, self.radii, [division_point, phi1])
 
             box1.segments[0] = self.segments[0]
             box2.segments[2] = self.segments[2]
-            box1.segments[2]._reversePath = box2.segments[0]
-            box2.segments[0]._reversePath = box1.segments[2]
+            box1.segments[2]._reverse_path = box2.segments[0]
+            box2.segments[0]._reverse_path = box1.segments[2]
+
+        else:
+            raise ValueError("axis must be 'r' or 'phi'")
 
         for box in [box1, box2]:
             box._created_by_subdivision_axis = axis
-            box._parentBox = self
-        self._childBoxes = [box1, box2]
+            box._parent = self
+        self._children = [box1, box2]
 
         return box1, box2