+++ changelog.md
@@ -1,6 +1,6 @@
 ## Changelog
 ### Unreleased
-- Drop support for Python 2
+- Drop support for Python 2. Cxroots now requires python 3.8 or later
 - cxroots now logs rootfinding progress which can be accessed using the standard library's [logging module](https://docs.python.org/3/library/logging.html). See the [documentation](https://rparini.github.io/cxroots/logging.html) for examples.
 - The `verbose` argument has been removed from some functions. It still exists for the `Contour.roots` method and `find_roots` function but it will now create a progress bar, using [rich](https://github.com/Textualize/rich), rather than printing debugging information to the console.
 - Use [Black](https://github.com/psf/black) formatting and added pre-commit hook
@@ -19,6 +19,7 @@
 - Changes default `root_tol` to 0 for `secant`, `newton` and `muller` functions
 - Removed `return_animation` argument from `demo_find_roots` function and `Contour.demo_roots` method. Instead, the `demo_roots_animation` function or `Contour.demo_roots_animation` method can be used to get a `animation.FuncAnimation` object that would animate the rootfinding process without displaying it.
 - Change starting points for muller's method used when root refining to be complex, to guard against the iterations keeping to the real line.
+- Rename `RootResult.original_contour` attribute to `contour`
 - The `Contour._size_plot` method was renamed to `Contour.size_plot` and given a docstring
 
 ### 1.1.11 - 22/Dec/2021